refactor(devops): Simplify cache and dev script all use ts code from src folder
