+++ scripts/buildCache.sh
@@ -3,8 +3,10 @@
 # Compile contracts
 yarn hardhat compile
 
-# Disable isolatedModules in tsconfig
+# Disable isolatedModules and use commonjs in tsconfig
 contents="$(jq '.compilerOptions.isolatedModules = false' tsconfig.json)" && \
+echo "${contents}" > tsconfig.json
+contents="$(jq '.compilerOptions.module = "commonjs"' tsconfig.json)" && \
 echo "${contents}" > tsconfig.json
 
 # Run build cache
@@ -12,6 +14,8 @@ yarn hardhat run --network rinkeby scripts/buildCache.ts
 yarn hardhat run --network arbitrum-testnet-v5 scripts/buildCache.ts
 sleep 1
 
-# Enable isolatedModules in tsconfig
+# Enable isolatedModules and use esnext as module in tsconfig
 contents="$(jq '.compilerOptions.isolatedModules = true' tsconfig.json)" && \
+echo "${contents}" > tsconfig.json
+contents="$(jq '.compilerOptions.module = "esnext"' tsconfig.json)" && \
 echo "${contents}" > tsconfig.json
